requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==3.1.5
django-cors-headers==3.6.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.8
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.10
ptyprocess==0.7.0
Pygments==2.7.3
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
whitenoise==5.2.0
